+++ html_stream/render_html_response.ts
@@ -11,13 +11,15 @@ import { renderHtmlBody } from "./render_html_body.ts";
  */
 export function renderHtmlResponse(
   node: HtmlNode,
-  options?: RenderOptions<HtmlToken>,
+  options?: RenderOptions<HtmlToken> & { headers?: HeadersInit },
 ): Response {
+  const headers = new Headers(options?.headers);
+  if (!headers.has("Content-Type")) {
+    headers.set("Content-Type", "text/html");
+  }
   return new Response(renderHtmlBody(node, options), {
     status: 200,
     statusText: "OK",
-    headers: {
-      "Content-Type": "text/html",
-    },
+    headers,
   });
 }