+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0]
+
+### Changed
+
+- [@http/token-stream] fix `deferralHandler` to emit deferred content in the
+  order it resolves rather than the order it was added.
+- [@http/html-stream] support passing headers to `renderHtmlResponse`.
+
 ## [0.5.0]
 
 ## BREAKING CHANGES